Added option for delayed key functionality
If enabled, a device that does not have the zad_BlockGeneric keyword attached is equipped will store a random time before it can be unequipped.
Devices that do not override the default RemoveDeviceWithKey() and DeviceMenuRemoveWithoutKey() methods will trigger new code to block the removal if the time has not yet been reached.